Add instrumentation logging to LogAnalyticsSendBufferTask and retry to LogAnalyticsClient
Finding the cause of errors sending data to Log Analytics has been challenging. This change
adds additional instrumentation and exception cause logging for errors occurring in
LogAnalyticsClient.send(...).  Additionally, a default retry policy (3 retries) is added
to the HttpClient that is used by LogAnalyticsClient, as well as an extended default
timeout policy.